refactor: separate service logic from presentation layer
