+++ src/main/kotlin/com/kroffle/knitting/controller/router/design/DesignRouter.kt
@@ -1,6 +1,6 @@
-package com.kroffle.knitting.controller.design
+package com.kroffle.knitting.controller.router.design
 
-import com.kroffle.knitting.usecase.design.DesignHandler
+import com.kroffle.knitting.controller.handler.design.DesignHandler
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.reactive.function.server.RequestPredicates.path
@@ -8,13 +8,13 @@ import org.springframework.web.reactive.function.server.RouterFunctions.nest
 import org.springframework.web.reactive.function.server.router
 
 @Configuration
-class DesignsRouter(private val handler: DesignHandler) {
+class DesignRouter(private val handler: DesignHandler) {
     @Bean
-    fun designsRouterFunction() = nest(
-        path("/designs"),
+    fun designRouterFunction() = nest(
+        path("/design"),
         router {
             listOf(
-                GET("/", handler::getAll),
+                POST("/", handler::createDesign),
             )
         }
     )
