+++ src/main/kotlin/com/kroffle/knitting/controller/router/design/DesignsRouter.kt
@@ -1,6 +1,6 @@
-package com.kroffle.knitting.controller.ping
+package com.kroffle.knitting.controller.router.design
 
-import com.kroffle.knitting.usecase.ping.PingHandler
+import com.kroffle.knitting.controller.handler.design.DesignHandler
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.reactive.function.server.RequestPredicates.path
@@ -8,12 +8,14 @@ import org.springframework.web.reactive.function.server.RouterFunctions.nest
 import org.springframework.web.reactive.function.server.router
 
 @Configuration
-class PingRouter(private val handler: PingHandler) {
+class DesignsRouter(private val handler: DesignHandler) {
     @Bean
-    fun pingRouterFunction() = nest(
-        path("/ping"),
+    fun designsRouterFunction() = nest(
+        path("/designs"),
         router {
-            listOf(GET("/", handler::get))
+            listOf(
+                GET("/", handler::getAll),
+            )
         }
     )
 }
