+++ src/main/kotlin/com/kroffle/knitting/controller/router/ping/PingRouter.kt
@@ -1,6 +1,6 @@
-package com.kroffle.knitting.controller.design
+package com.kroffle.knitting.controller.router.ping
 
-import com.kroffle.knitting.usecase.design.DesignHandler
+import com.kroffle.knitting.controller.handler.ping.PingHandler
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.reactive.function.server.RequestPredicates.path
@@ -8,14 +8,12 @@ import org.springframework.web.reactive.function.server.RouterFunctions.nest
 import org.springframework.web.reactive.function.server.router
 
 @Configuration
-class DesignRouter(private val handler: DesignHandler) {
+class PingRouter(private val handler: PingHandler) {
     @Bean
-    fun designRouterFunction() = nest(
-        path("/design"),
+    fun pingRouterFunction() = nest(
+        path("/ping"),
         router {
-            listOf(
-                POST("/", handler::createDesign),
-            )
+            listOf(GET("/", handler::get))
         }
     )
 }
