+++ src/main/kotlin/com/kroffle/knitting/infra/oauth/GoogleOauthHelperImpl.kt
@@ -1,8 +1,8 @@
 package com.kroffle.knitting.infra.oauth
 
+import com.kroffle.knitting.controller.handler.auth.GoogleLogInHandler
 import com.kroffle.knitting.infra.properties.SelfProperties
 import com.kroffle.knitting.infra.properties.WebApplicationProperties
-import com.kroffle.knitting.usecase.auth.GoogleLogInHandler
 import org.springframework.web.util.UriComponentsBuilder
 
 class GoogleOauthHelperImpl(
